requirements.txt
pandas==1.4.4
selenium==4.4.3
psycopg2-binary==2.9.5